# src/Coral.Api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Coral",
    "version": "v1"
  },
  "paths": {
...
    "/api/Onboarding/listDirectories": {
      "get": {
        "tags": [
          "Onboarding"
        ],
        "operationId": "DirectoriesInPath",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Onboarding/register": {
      "post": {
        "tags": [
          "Onboarding"
        ],
        "operationId": "RegisterMusicLibrary",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MusicLibrary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicLibrary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicLibrary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioFile": {
        "required": [
          "audioMetadata",
          "audioMetadataId",
          "createdAt",
          "filePath",
          "fileSizeInBytes",
          "id",
          "library",
          "libraryId",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "filePath": {
            "type": "string"
          },
          "fileSizeInBytes": {
            "type": "number",
            "format": "double"
          },
          "audioMetadataId": {
            "type": "string",
            "format": "uuid"
          },
          "audioMetadata": {
            "$ref": "#/components/schemas/AudioMetadata"
          },
          "libraryId": {
            "type": "string",
            "format": "uuid"
          },
          "library": {
            "$ref": "#/components/schemas/MusicLibrary"
          }
        },
        "additionalProperties": false
      },
      "AudioMetadata": {
        "required": [
          "bitrate",
          "codec",
          "createdAt",
          "id",
          "sampleRate",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "bitrate": {
            "type": "integer",
            "format": "int32"
          },
          "bitDepth": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sampleRate": {
            "type": "number",
            "format": "double"
          },
          "channels": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "codec": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "MusicLibrary": {
        "required": [
          "audioFiles",
          "createdAt",
          "id",
          "lastScan",
          "libraryPath",
          "updatedAt",
          "watchForChanges"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "libraryPath": {
            "type": "string"
          },
          "lastScan": {
            "type": "string",
            "format": "date-time"
          },
          "watchForChanges": {
            "type": "boolean"
          },
          "audioFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudioFile"
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}